move convert_c4
