+++ .gitignore
@@ -1,3 +1,9 @@
+# Data
+my-copy-c4/
+
+# WandB
+*wandb/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
