+++ examples/llm/tests/test_c4_data_prep_script.py
@@ -2,13 +2,10 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import os
-import pathlib
 import shutil
-import sys
 from argparse import Namespace
 
-sys.path.append(str(pathlib.Path(__file__).parent.parent.parent / 'scripts'))
-from convert_c4 import main  # type: ignore
+from examples.common.convert_c4 import main
 
 
 def test_download_script_from_api():
@@ -30,7 +27,6 @@ def test_download_script_from_cmdline():
     path = os.path.join(os.getcwd(), 'my-copy-c4-2')
     shutil.rmtree(path, ignore_errors=True)
     os.system(
-        'python ../scripts/convert_c4.py --out_root ./my-copy-c4-2 --splits val'
-    )
+        'python ../common/convert_c4.py --out_root ./my-copy-c4-2 --splits val')
     assert os.path.exists(path)
     shutil.rmtree(path, ignore_errors=False)
