--- scripts/__init__.py
@@ -1,2 +1,0 @@
-# Copyright 2022 MosaicML Examples authors
-# SPDX-License-Identifier: Apache-2.0
